TESTS: Move get_call_output() to util.py
This function will be reused outside of test_sssctl.py.

Related:
https://pagure.io/SSSD/sssd/issue/3658
