+++ src/tests/intg/util.py
@@ -78,3 +78,10 @@ def restore_envvar_file(name):
     path = os.environ[name]
     backup_path = path + ".bak"
     os.rename(backup_path, path)
+
+
+def get_call_output(cmd):
+    process = subprocess.Popen(cmd, stdout=subprocess.PIPE,
+                               stderr=subprocess.PIPE)
+    output, ret = process.communicate()
+    return output.decode('utf-8')
